--- dataloader.py
@@ -1,134 +1,0 @@
-from sklearn.preprocessing import LabelEncoder
-import numpy as np
-import torch
-
-class Preprocess:
-    def __init__(self, args, le):
-        self.args = args
-        self.train_data = None
-        self.test_data = None
-        self.le = le
-
-    def get_test_data(self):
-        return self.test_data
-        
-    def __preprocessing(self, df, is_train = True):
-        cate_cols = ['assessmentItemID', 'testId', 'KnowledgeTag']
-
-        for col in cate_cols:
-            
-            le = LabelEncoder()
-            
-            le.classes_ = self.le[col]
-            
-            df[col] = df[col].apply(lambda x: x if str(x) in le.classes_ else 'unknown')
-
-            #모든 컬럼이 범주형이라고 가정
-            df[col]= df[col].astype(str)
-            
-            test = le.transform(df[col])
-            df[col] = test
-            
-        return df
-    
-
-    def __feature_engineering(self, df):
-        #TODO
-        return df
-
-    def load_data(self, df):
-        
-        df = self.__feature_engineering(df)
-        df = self.__preprocessing(df)
-           
-        columns = ['userID', 'assessmentItemID', 'testId', 'answerCode', 'KnowledgeTag']
-        group = df[columns].groupby('userID').apply(
-                lambda r: (
-                    r['testId'].values, 
-                    r['assessmentItemID'].values,
-                    r['KnowledgeTag'].values,
-                    r['answerCode'].values
-                )
-            )
-
-        return group.values
-
-    def load_test_data(self, data):
-        self.test_data = self.load_data(data)
-
-
-class DKTDataset(torch.utils.data.Dataset):
-    def __init__(self, data, args):
-        self.data = data
-        self.args = args
-
-    def __getitem__(self, index):
-        row = self.data[index]
-
-        # 각 data의 sequence length
-        seq_len = len(row[0])
-
-        test, question, tag, correct = row[0], row[1], row[2], row[3]
-        
-
-        cate_cols = [test, question, tag, correct]
-
-        # max seq len을 고려하여서 이보다 길면 자르고 아닐 경우 그대로 냅둔다
-        if seq_len > self.args.max_seq_len:
-            for i, col in enumerate(cate_cols):
-                cate_cols[i] = col[-self.args.max_seq_len:]
-            mask = np.ones(self.args.max_seq_len, dtype=np.int16)
-        else:
-            mask = np.zeros(self.args.max_seq_len, dtype=np.int16)
-            mask[-seq_len:] = 1
-
-        # mask도 columns 목록에 포함시킴
-        cate_cols.append(mask)
-
-        # np.array -> torch.tensor 형변환
-        for i, col in enumerate(cate_cols):
-            cate_cols[i] = torch.tensor(col)
-
-        return cate_cols
-
-    def __len__(self):
-        return len(self.data)
-
-
-from torch.nn.utils.rnn import pad_sequence
-
-def collate(batch):
-    col_n = len(batch[0])
-    col_list = [[] for _ in range(col_n)]
-    max_seq_len = len(batch[0][-1])
-
-        
-    # batch의 값들을 각 column끼리 그룹화
-    for row in batch:
-        for i, col in enumerate(row):
-            pre_padded = torch.zeros(max_seq_len)
-            pre_padded[-len(col):] = col
-            col_list[i].append(pre_padded)
-
-
-    for i, _ in enumerate(col_list):
-        col_list[i] =torch.stack(col_list[i])
-    
-    return tuple(col_list)
-
-
-def get_loaders(args, train, valid):
-
-    pin_memory = False
-    train_loader, valid_loader = None, None
-    
-    if train is not None:
-        trainset = DKTDataset(train, args)
-        train_loader = torch.utils.data.DataLoader(trainset, num_workers=args.num_workers, shuffle=True,
-                            batch_size=args.batch_size, pin_memory=pin_memory, collate_fn=collate)
-    if valid is not None:
-        valset = DKTDataset(valid, args)
-        valid_loader = torch.utils.data.DataLoader(valset, num_workers=0, shuffle=False,
-                            batch_size=args.batch_size, pin_memory=pin_memory, collate_fn=collate)
-
-    return train_loader, valid_loader
