+++ dkt/dataloader.py
@@ -1,99 +1,46 @@
-import os
-from datetime import datetime
-import time
-import tqdm
-import pandas as pd
-import random
 from sklearn.preprocessing import LabelEncoder
 import numpy as np
 import torch
 
 class Preprocess:
-    def __init__(self,args):
+    def __init__(self, args, le):
         self.args = args
         self.train_data = None
         self.test_data = None
-        
-
-    def get_train_data(self):
-        return self.train_data
+        self.le = le
 
     def get_test_data(self):
         return self.test_data
-
-    def split_data(self, data, ratio=0.7, shuffle=True, seed=0):
-        """
-        split data into two parts with a given ratio.
-        """
-        if shuffle:
-            random.seed(seed) # fix to default seed 0
-            random.shuffle(data)
-
-        size = int(len(data) * ratio)
-        data_1 = data[:size]
-        data_2 = data[size:]
-
-        return data_1, data_2
-
-    def __save_labels(self, encoder, name):
-        le_path = os.path.join(self.args.asset_dir, name + '_classes.npy')
-        np.save(le_path, encoder.classes_)
-
+        
     def __preprocessing(self, df, is_train = True):
         cate_cols = ['assessmentItemID', 'testId', 'KnowledgeTag']
 
-        if not os.path.exists(self.args.asset_dir):
-            os.makedirs(self.args.asset_dir)
-            
         for col in cate_cols:
             
-            
             le = LabelEncoder()
-            if is_train:
-                #For UNKNOWN class
-                a = df[col].unique().tolist() + ['unknown']
-                le.fit(a)
-                self.__save_labels(le, col)
-            else:
-                label_path = os.path.join(self.args.asset_dir,col+'_classes.npy')
-                le.classes_ = np.load(label_path)
-                
-                df[col] = df[col].apply(lambda x: x if x in le.classes_ else 'unknown')
+            
+            le.classes_ = self.le[col]
+            
+            df[col] = df[col].apply(lambda x: x if str(x) in le.classes_ else 'unknown')
 
             #모든 컬럼이 범주형이라고 가정
             df[col]= df[col].astype(str)
+            
             test = le.transform(df[col])
             df[col] = test
             
-
-        def convert_time(s):
-            timestamp = time.mktime(datetime.strptime(s, '%Y-%m-%d %H:%M:%S').timetuple())
-            return int(timestamp)
-
-        df['Timestamp'] = df['Timestamp'].apply(convert_time)
-        
         return df
+    
 
     def __feature_engineering(self, df):
         #TODO
         return df
 
-    def load_data_from_file(self, file_name, is_train=True):
-        csv_file_path = os.path.join(self.args.data_dir, file_name)
-        df = pd.read_csv(csv_file_path)#, nrows=100000)
-        df = self.__feature_engineering(df)
-        df = self.__preprocessing(df, is_train)
-
-        # 추후 feature를 embedding할 시에 embedding_layer의 input 크기를 결정할때 사용
-
-                
-        self.args.n_questions = len(np.load(os.path.join(self.args.asset_dir,'assessmentItemID_classes.npy')))
-        self.args.n_test = len(np.load(os.path.join(self.args.asset_dir,'testId_classes.npy')))
-        self.args.n_tag = len(np.load(os.path.join(self.args.asset_dir,'KnowledgeTag_classes.npy')))
+    def load_data(self, df):
         
-
-
-        df = df.sort_values(by=['userID','Timestamp'], axis=0)
+        df = self.__feature_engineering(df)
+        df = self.__preprocessing(df)
+           
         columns = ['userID', 'assessmentItemID', 'testId', 'answerCode', 'KnowledgeTag']
         group = df[columns].groupby('userID').apply(
                 lambda r: (
@@ -106,11 +53,8 @@ class Preprocess:
 
         return group.values
 
-    def load_train_data(self, file_name):
-        self.train_data = self.load_data_from_file(file_name)
-
-    def load_test_data(self, file_name):
-        self.test_data = self.load_data_from_file(file_name, is_train= False)
+    def load_test_data(self, data):
+        self.test_data = self.load_data(data)
 
 
 class DKTDataset(torch.utils.data.Dataset):
@@ -184,7 +128,7 @@ def get_loaders(args, train, valid):
                             batch_size=args.batch_size, pin_memory=pin_memory, collate_fn=collate)
     if valid is not None:
         valset = DKTDataset(valid, args)
-        valid_loader = torch.utils.data.DataLoader(valset, num_workers=args.num_workers, shuffle=False,
+        valid_loader = torch.utils.data.DataLoader(valset, num_workers=0, shuffle=False,
                             batch_size=args.batch_size, pin_memory=pin_memory, collate_fn=collate)
 
     return train_loader, valid_loader
